refactor: move splitRawConfig into @cloudflare/config
splitRawConfig is the full-fidelity inverse of @cloudflare/config's
convertToWranglerConfig, so it belongs alongside the forward converter
rather than in autoconfig. Move split.ts (and its unit tests) into
@cloudflare/config, export splitRawConfig / NewConfigSplit from the
package index, and have autoconfig import from @cloudflare/config.

autoconfig's serializer tests stay in autoconfig (they exercise
serialize.ts, which remains there).
