+++ packages/config/src/index.ts
@@ -17,6 +17,8 @@ export {
 } from "./schema";
 export { generateTypes } from "./generate";
 export { convertToWranglerConfig } from "./convert";
+export { splitRawConfig } from "./split";
+export type { NewConfigSplit } from "./split";
 export { loadConfig, registerConfigHooks } from "./load";
 export { resolveWorkerDefinition } from "./worker-definition";
 export type { LoadConfigResult } from "./load";
